Revert "Commit to rollback JsonSubTypes impl"
This reverts commit bfd7350a
