+++ src/main/java/it/auties/whatsapp4j/model/WhatsappResponse.java
@@ -1,29 +1,32 @@
 package it.auties.whatsapp4j.model;
 
-import it.auties.whatsapp4j.response.model.NoResponse;
+import com.fasterxml.jackson.annotation.JsonSubTypes;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import it.auties.whatsapp4j.response.model.JsonListResponse;
+import it.auties.whatsapp4j.response.model.JsonResponse;
 import it.auties.whatsapp4j.response.model.Response;
+import lombok.*;
 import jakarta.validation.constraints.NotNull;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
 import lombok.experimental.Accessors;
+
+
+import java.io.IOException;
+import java.util.HashMap;
 
 @Accessors(chain = true,fluent = true)
 @Getter
 @Setter
 @EqualsAndHashCode(of = {"tag"})
 @ToString(of = "tag")
+@Builder
 public class WhatsappResponse<J extends WhatsappResponse<J>> {
-    @NotNull
     private String tag;
     private String description;
     private Response<?,?> data;
 
     public WhatsappResponse() {
-        //No config required
-        tag = "pong";
-        data = new NoResponse<>();
     }
 
     /**
@@ -31,10 +34,66 @@ public class WhatsappResponse<J extends WhatsappResponse<J>> {
      *
      * @param tag         the tag used for the request
      * @param description a nullable String that describes how to categorize the data that is object holds
+     * @param data        the data that this object holds
      */
-    public WhatsappResponse(@NotNull String tag, String description) {
+    public WhatsappResponse(@NotNull String tag, String description, @NotNull Response<?,?> data) {
         this.tag = tag;
         this.description = description;
+        this.data = data;
+    }
+
+    /**
+     * An instance of Jackson used to deserialize JSON Strings
+     */
+    private static final ObjectMapper JACKSON = new ObjectMapper().configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+
+
+    /**
+     * Constructs a new instance of WhatsappResponse from a json string
+     *
+     * @param parse the json string to parse
+     * @return a new instance of WhatsappResponse with the above characteristics
+     * @throws IllegalArgumentException if {@code parse} cannot be parsed
+     */
+    public @NotNull WhatsappResponse<J> fromJson(@NotNull String parse) {
+        return new WhatsappResponse<>("tag","description",null);
+
+        /*try {
+            var split = parse.split(",", 2);
+            if (split.length != 2 && parse.startsWith("!")) {
+                return new WhatsappResponse(parse, "pong", new JsonResponse(new HashMap<>()));
+            }
+
+            var response = WhatsappResponse.tag(split[0]);
+            var content = parseContent(split[1], 0);
+            if (content.isEmpty()) {
+                return response.data(new JsonResponse(new HashMap<>())).build();
+            }
+
+            var jsonNode = JACKSON.readTree(content);
+            if (!jsonNode.isArray()) {
+                return response.data(new JsonResponse(JACKSON.readerFor(new TypeReference<>() {
+                }).readValue(jsonNode))).build();
+            }
+
+            var possibleMap = jsonNode.get(1);
+            if (possibleMap == null) {
+                return response.data(new JsonListResponse(JACKSON.readerFor(new TypeReference<>() {
+                }).readValue(jsonNode))).build();
+            }
+
+            var possibleMapContent = possibleMap.toString();
+            if (!possibleMapContent.startsWith("{") || !possibleMapContent.endsWith("}")) {
+                return response.data(new JsonListResponse(JACKSON.readerFor(new TypeReference<>() {
+                }).readValue(jsonNode))).build();
+            }
+
+            return response.description(jsonNode.get(0).textValue()).data(new JsonResponse(JACKSON.readerFor(new TypeReference<>() {
+            }).readValue(possibleMap))).build();
+        } catch (IOException ex) {
+            throw new IllegalArgumentException("WhatsappAPI: Cannot deserialize %s into a WhatsappResponse".formatted(parse));
+        }*/
+
     }
 
     private static @NotNull String parseContent(@NotNull String content, int index) {
