+++ src/main/java/it/auties/whatsapp4j/request/impl/LogOutRequest.java
@@ -1,6 +1,5 @@
 package it.auties.whatsapp4j.request.impl;
 
-import com.fasterxml.jackson.annotation.JsonTypeName;
 import it.auties.whatsapp4j.api.WhatsappConfiguration;
 import it.auties.whatsapp4j.request.model.JsonRequest;
 import it.auties.whatsapp4j.response.impl.DiscardResponse;
@@ -11,7 +10,6 @@ import java.util.List;
 /**
  * A JSON request used to signal to WhatsappWeb's WebSocket that this session's keys should be invalidated and the session closed
  */
-@JsonTypeName("logoutrequest")
 public class LogOutRequest extends JsonRequest<LogOutRequest,DiscardResponse> {
     public LogOutRequest(@NotNull WhatsappConfiguration configuration) {
         super(configuration);
