+++ src/main/java/it/auties/whatsapp4j/request/impl/NodeRequest.java
@@ -1,6 +1,5 @@
 package it.auties.whatsapp4j.request.impl;
 
-import com.fasterxml.jackson.annotation.JsonTypeName;
 import it.auties.whatsapp4j.api.WhatsappConfiguration;
 import it.auties.whatsapp4j.model.WhatsappNode;
 import it.auties.whatsapp4j.request.model.BinaryRequest;
@@ -10,7 +9,6 @@ import jakarta.validation.constraints.NotNull;
 /**
  * A binary request used to transmit a {@link WhatsappNode} to WhatsappWeb's WebSocket
  */
-@JsonTypeName("noderequest")
 public class NodeRequest<M extends ResponseModel<M>> extends BinaryRequest<M> {
 
     private final @NotNull WhatsappNode node;
