+++ src/main/java/it/auties/whatsapp4j/response/impl/DiscardResponse.java
@@ -25,4 +25,6 @@ public final class DiscardResponse implements JsonResponseModel<DiscardResponse>
     public String toString() {
         return "DiscardResponse[]";
     }
+
+
 }
