+++ src/main/java/it/auties/whatsapp4j/response/model/Response.java
@@ -21,7 +21,6 @@ import jakarta.validation.constraints.NotNull;
         }
 )
 public interface Response<J extends Response<J, T>, T extends ResponseModel<T>> {
-
     /**
      * Converts this object to a ResponseModel
      *
